+++ CHANGELOG.md
@@ -7,6 +7,10 @@ This projects adheres to [Semantic Versioning](http://semver.org/) and [Keep a C
 ## [Unreleased][unreleased]
 -
 
+## [2.0.3] - 2018-12-12
+- Use issuer field from core gateway.
+- Updated deprecated function calls.
+
 ## [2.0.2] - 2018-05-14
 - Switched to PHP namespaces.
 
@@ -35,7 +39,8 @@ This projects adheres to [Semantic Versioning](http://semver.org/) and [Keep a C
 ## 1.0.0 - 2015-01-19
 - First release.
 
-[unreleased]: https://github.com/wp-pay-gateways/multisafepay/compare/2.0.2...HEAD
+[unreleased]: https://github.com/wp-pay-gateways/multisafepay/compare/2.0.3...HEAD
+[2.0.3]: https://github.com/wp-pay-gateways/multisafepay/compare/2.0.2...2.0.3
 [2.0.2]: https://github.com/wp-pay-gateways/multisafepay/compare/2.0.1...2.0.2
 [2.0.1]: https://github.com/wp-pay-gateways/multisafepay/compare/2.0.0...2.0.1
 [2.0.0]: https://github.com/wp-pay-gateways/multisafepay/compare/1.1.4...2.0.0